+++ c/isotope.h
@@ -39,28 +39,28 @@
 
 void
 iso_get_isotope_scattering_strength(double *gamma,
-                                    const size_t grid_point,
+                                    const long grid_point,
                                     const double *mass_variances,
                                     const double *frequencies,
                                     const lapack_complex_double *eigenvectors,
-                                    const size_t num_grid_points,
+                                    const long num_grid_points,
                                     const int *band_indices,
-                                    const size_t num_band,
-                                    const size_t num_band0,
+                                    const long num_band,
+                                    const long num_band0,
                                     const double sigma,
                                     const double cutoff_frequency);
 void iso_get_thm_isotope_scattering_strength
 (double *gamma,
- const size_t grid_point,
- const size_t *ir_grid_points,
- const int *weights,
+ const long grid_point,
+ const long *ir_grid_points,
+ const long *weights,
  const double *mass_variances,
  const double *frequencies,
  const lapack_complex_double *eigenvectors,
- const size_t num_grid_points,
+ const long num_grid_points,
  const int *band_indices,
- const size_t num_band,
- const size_t num_band0,
+ const long num_band,
+ const long num_band0,
  const double *integration_weights,
  const double cutoff_frequency);
 #endif
