+++ c/triplet_kpoint.h
@@ -41,20 +41,20 @@
 #include "mathfunc.h"
 #include "triplet.h"
 
-size_t tpk_get_ir_triplets_at_q(size_t *map_triplets,
-                                size_t *map_q,
-                                int (*grid_address)[3],
-                                const int grid_point,
-                                const int mesh[3],
-                                const int is_time_reversal,
-                                const MatINT * rotations,
-                                const int swappable);
-size_t tpk_get_BZ_triplets_at_q(size_t (*triplets)[3],
-                                const size_t grid_point,
-                                TPLCONST int (*bz_grid_address)[3],
-                                const size_t *bz_map,
-                                const size_t *map_triplets,
-                                const size_t num_map_triplets,
-                                const int mesh[3]);
+long tpk_get_ir_triplets_at_q(long *map_triplets,
+                              long *map_q,
+                              int (*grid_address)[3],
+                              const int grid_point,
+                              const int mesh[3],
+                              const int is_time_reversal,
+                              const MatINT * rotations,
+                              const int swappable);
+long tpk_get_BZ_triplets_at_q(long (*triplets)[3],
+                              const long grid_point,
+                              TPLCONST int (*bz_grid_address)[3],
+                              const long *bz_map,
+                              const long *map_triplets,
+                              const long num_map_triplets,
+                              const int mesh[3]);
 
 #endif
